separate class for details endpoints
