+++ Api/SettingsInterface.php
@@ -16,18 +16,4 @@ interface SettingsInterface {
      * @return string
      */
     public function showStatus($websiteId = 0);
-
-    /**
-     * GET for order details
-     * @param string $orderId
-     * @return string
-     */
-    public function orderDetails($orderId);
-
-    /**
-     * POST for product details
-     * @param string $productId
-     * @return string
-     */
-    public function productDetails($productId);
 }
