+++ etc/di.xml
@@ -15,4 +15,6 @@
     </type>
     <preference for="Drip\Connect\Api\SettingsInterface"
             type="Drip\Connect\Api\Settings" />
+    <preference for="Drip\Connect\Api\ResourceDetailsInterface"
+            type="Drip\Connect\Api\ResourceDetails" />
 </config>
