+++ etc/webapi.xml
@@ -13,13 +13,13 @@
         </resources>
     </route>
     <route method="GET" url="/V1/drip/order/:orderId">
-        <service class="Drip\Connect\Api\SettingsInterface" method="orderDetails"/>
+        <service class="Drip\Connect\Api\ResourceDetailsInterface" method="orderDetails"/>
         <resources>
             <resource ref="Drip::settings"/>
         </resources>
     </route>
     <route method="GET" url="/V1/drip/product/:productId">
-        <service class="Drip\Connect\Api\SettingsInterface" method="productDetails"/>
+        <service class="Drip\Connect\Api\ResourceDetailsInterface" method="productDetails"/>
         <resources>
             <resource ref="Drip::settings"/>
         </resources>
